+++ jobbergate-api/tests/conftest.py
@@ -10,10 +10,13 @@ from contextlib import asynccontextmanager
 from textwrap import dedent
 from unittest.mock import patch
 
+import aio_pika
 import pytest
+import socket
 from fastapi import status
 from httpx import AsyncClient, Response
 from sqlalchemy.ext.asyncio import AsyncSession
+
 
 from jobbergate_api.apps.dependencies import get_bucket_name, get_bucket_url, s3_bucket, service_factory
 from jobbergate_api.apps.models import Base